Multiple goals

+++ app/src/main/java/com/example/appdev/database/entities/RecurringCostEntity.kt
@@ -9,13 +9,13 @@ import androidx.room.PrimaryKey
         ForeignKey(entity = GoalEntity::class, parentColumns = ["goal_id"], childColumns = ["goal_id"], onDelete = ForeignKey.CASCADE),
         ForeignKey(entity = UserEntity::class, parentColumns = ["user_id"], childColumns = ["user_id"], onDelete = ForeignKey.CASCADE)
     ]
-    )
+)
 data class RecurringCostEntity (
     @PrimaryKey(autoGenerate = true) val cost_id: Int = 0,
     val goal_id : Int,
     val user_id : Int,
     val title : String,
-    val amount : Float,
+    val amount : Double,
     val currency : String,
     val frequency : String
 )
